Add logging to main.py
Logging is only set up once in main.py and affects all submodules.

In each module that you want to use logging:

	import logging

	# Get a logger for the module immediately after imports, before other code.
	log = logging.getLogger(__name__)

	# Somewhere in your code
	log.debug("Hello devs")
	log.info("Saved profile: %s", profile.name)

Use printf-style %s, %d etc for placeholders, and pass args to the logger so
that they are only interpolated when the message is actually logged.

	# Correct
	log.warning("Some variables: %d %s %r", count, name, obj)
	# Wrong
	log.warning("Some variables: %d %s %r" % (count, name, obj))
	log.warning(f"Some variables: {count} {name} {obj})

This is only for logging. Use f-strings elsewhere as usual.

https://docs.python.org/3/library/logging.html